+++ magento/loaders/productDetailsPage.ts
@@ -1,12 +1,12 @@
 import type { ListItem, ProductDetailsPage } from "../../commerce/types.ts";
-import { RequestPathname } from "../functions/requestToPathname.ts";
+import { RequestURLParam } from "../../website/functions/requestToParam.ts";
 import { AppContext } from "../mod.ts";
 import { URL_KEY } from "../utils/constants.ts";
 import stringifySearchCriteria from "../utils/stringifySearchCriteria.ts";
 import { toBreadcrumbList, toProduct, toSeo } from "../utils/transform.ts";
 
 export interface Props {
-  slug: RequestPathname;
+  slug: RequestURLParam;
   isBreadcrumbProductName?: boolean;
 }
