+++ magento/manifest.gen.ts
@@ -6,15 +6,11 @@ import * as $$$$$$$$$0 from "./actions/cart/addItem.ts";
 import * as $$$$$$$$$1 from "./actions/cart/removeItem.ts";
 import * as $$$$$$$$$2 from "./actions/cart/simulation.ts";
 import * as $$$$$$$$$3 from "./actions/cart/updateItem.ts";
-import * as $0 from "./functions/requestToPathname.ts";
 import * as $$$0 from "./loaders/cart.ts";
 import * as $$$1 from "./loaders/productDetailsPage.ts";
 import * as $$$2 from "./loaders/user.ts";
 
 const manifest = {
-  "functions": {
-    "magento/functions/requestToPathname.ts": $0,
-  },
   "loaders": {
     "magento/loaders/cart.ts": $$$0,
     "magento/loaders/productDetailsPage.ts": $$$1,
